requirements.txt
django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-cloudinary-storage==0.3.0
cloudinary==1.40.0
django-jazzmin==3.0.0
django-filter==24.2
psycopg2-binary==2.9.9
razorpay==1.4.1
requests==2.32.3
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.7.0
Pillow==10.4.0
django-ratelimit==4.1.0
dj-database-url==2.2.0
setuptools>=65.0.0
